Changed the text field to parse the value field as a double.

+++ src/presentation/AddFeatureHistoryDrawer.java
@@ -58,7 +58,7 @@ public class AddFeatureHistoryDrawer extends BaseFeatureHistoryDrawer
 			formatter = new SimpleDateFormat("dd/MM/yyyy");
 			date = formatter.parse(dateTime.getDay() + "/" + dateTime.getMonth() + "/" + dateTime.getYear());
 			
-			history = new FeatureHistory(feature, trackedService, Integer.parseInt(txtValue.getText()), date, txtNotes.getText());
+			history = new FeatureHistory(feature, trackedService, Double.parseDouble(txtValue.getText()), date, txtNotes.getText());
 			
 			if(processFeatureHistory.insert(history))
 				backToPreviousScreen();
